+++ src/main/java/hotspot/admin/outbox/notificationOutbox/infrastructure/NotificationOutboxEventJpaRepository.java
@@ -1,10 +1,10 @@
-package hotspot.admin.outbox.infrastructure;
+package hotspot.admin.outbox.notificationOutbox.infrastructure;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import hotspot.admin.outbox.infrastructure.entity.NotificationOutboxEventEntity;
+import hotspot.admin.outbox.notificationOutbox.infrastructure.entity.NotificationOutboxEventEntity;
 
 public interface NotificationOutboxEventJpaRepository extends JpaRepository<NotificationOutboxEventEntity, UUID> {
 }
