+++ src/test/java/hotspot/admin/family/outbox/FamilyRequestOutboxPublisherTest.java
@@ -24,7 +24,7 @@ import hotspot.admin.family.domain.ApplyType;
 import hotspot.admin.family.domain.FamilyApply;
 import hotspot.admin.family.domain.FamilyApplyStatus;
 import hotspot.admin.family.outbox.dto.FamilyRequestAlertEvent;
-import hotspot.admin.outbox.service.NotificationOutboxEventAppender;
+import hotspot.admin.outbox.notificationOutbox.service.NotificationOutboxEventAppender;
 
 @ExtendWith(MockitoExtension.class)
 class FamilyRequestOutboxPublisherTest {
